# Concurrently add partitions to Hive Metastore