… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveConfig.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestingThriftHiveMetastoreBuilder.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/metastore/thrift/TestThriftMetastoreConfig.java