<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

Increase the performance of the finishing step when inserting into the newly created partitioned table.

## Benchmarks

### For Glue:
Total partitions: 2000+
Total rows: 600M

Before:
Writing time: 2:29 mins
Finishing time: 4:32 mins
Total time: 7:01 mins

After:
Writing time: 2:27 mins
Finishing time: 1:01 mins
Total time: 3:28 mins

### For hive thrift (local setup):
Total partitions: 100+
Total rows: 1M

Before:
Writing time: 20 secs
Finishing time: 1:50 mins
Total time: 2:10 mins

After:
Writing time: 20 secs
Finishing time: 17.50 secs
Total time: 37.50 secs

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
( ) Release notes are required, with the following suggested text:

```markdown
# Section
* Fix some things. ({issue}`issuenumber`)
```